sherpa lib improvement/fix: for functions that return only an error, return a "null" result if there was no error
before, we would respond with an empty array as result.  that was
incorrect.  properly noone was actually looking at the result fo
such functions, but still.  this also changes to always return a
"result" field in the response.  of course, "error" still takes
precedence if set (callers (should) look at errors first).