parse parameters to function calls strictly by default
and allow user to opt out using the new HandlerOpts, with LaxParameterParsing.
this needs a new Go, because of encoding/json's Decoder.DisallowUnknownFields.

this changes the signature of NewHandler(). if you've passed in a
nil collector (last argument), your code will still compile (and
you get the default benefit of stricter parameter parsing).