turn sherpadoc into a library first
cmd/sherpadoc just calls the Main from the library.

this is a bit nasty.  the idea is that i want to use sherpadoc in
build scripts in other projects.  but with the new "go mod", i can't
vendor in code that isn't needed to build the project, i.e. actually
referenced. normally that's a good thing, and it's neat that with
"go mod -vendor" my vendor/ dir became much slimmer, but i want a
simple way to get the sherpadoc code included in my other project's
repo. because that gets me the benefits of "go mod", e.g. easy
upgrades.

you can't import a "main"-package in a program. it must be a library.
so this approach just turns all code that was previously "main"
into a library.  in my other projects - those that use sherpadoc -
i just write a tiny program that imports this sherpadoc library
code, and calls its Main function. that way "go mod" will find it
and vendor in the sherpadoc code.