move back the "library" code in sherpadoc/ into cmd/sherpadoc/
the reason it was moved before was to be able to get supporting
tools into a go modules-managed repository. tools with a "main"
package cannot normally be imported. but it can be done now, it's
in the go modules wiki page.  so remove the hack of putting all
main() code into a library.